requirements.txt
requests==2.31.0
google-cloud-storage==2.14.0
pydantic==2.6.1
pyyaml==6.0.1